add models for newly created hourly caggs and update queries to query from hourly caggs if real time data is requested

+++ app/models/classification_counts/daily_project_classification_count.rb
@@ -4,7 +4,7 @@ module ClassificationCounts
   class DailyProjectClassificationCount < ApplicationRecord
     self.table_name = 'daily_classification_count_and_time_per_project'
     attribute :classification_count, :integer
-    attribute :total_session_time, :integer
+    attribute :total_session_time, :float
 
     def readonly?
       true
